Move lock button to submissions label
